再生速度変更を app に持ってきた、けど playbackPosition どこに入れよ、DisplayTime か？

+++ app/src/main/java/io/github/takusan23/akaridroid/audiorender/AudioRender.kt
@@ -123,7 +123,7 @@ class AudioRender(
             // 戻れないので、InputStream を開き直す
             inputStream?.close()
             inputStream = outPcmFile.inputStream()
-            inputStream?.skip(seekBytePos.toLong())
+            inputStream?.skip(seekBytePos)
         }
     }
